refactor: arbitrum test.
